Release

+++ AutoEvent/DebugLogger.cs
@@ -1,17 +1,23 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
+#if EXILED
 using Exiled.API.Features;
+#else
+using LabApi.Features.Console;
+#endif
 
 namespace AutoEvent;
+
 public class DebugLogger
 {
     public static DebugLogger Singleton;
-    private string _filePath;
-    private static bool _loaded = false;
+    private static bool _loaded;
     public static bool Debug = false;
-    public static bool WriteDirectly = false;
-    private List<string> _debugLogs;
+    public static bool WriteDirectly;
+    private readonly List<string> _debugLogs;
+    private readonly string _filePath;
+
     public DebugLogger(bool writeDirectly)
     {
         Singleton = this;
@@ -19,65 +25,69 @@ public class DebugLogger
         _debugLogs = new List<string>();
         _loaded = true;
 
-        if (!Directory.Exists(AutoEvent.BaseConfigPath))
-        {
-            Directory.CreateDirectory(AutoEvent.BaseConfigPath);
-        }
+        if (!Directory.Exists(AutoEvent.BaseConfigPath)) Directory.CreateDirectory(AutoEvent.BaseConfigPath);
 
         try
         {
             _filePath = Path.Combine(AutoEvent.BaseConfigPath, "debug-output.log");
             if (WriteDirectly)
             {
-                DebugLogger.LogDebug($"Writing debug output directly to \"{_filePath}\"");
-                if (File.Exists(_filePath))
-                {
-                    File.Delete(_filePath);
-                }
+                LogDebug($"Writing debug output directly to \"{_filePath}\"");
+                if (File.Exists(_filePath)) File.Delete(_filePath);
 
                 File.Create(_filePath).Close();
             }
         }
         catch (Exception e)
         {
-            DebugLogger.LogDebug($"An error has occured while trying to create a debug log.", LogLevel.Warn, true);
-            DebugLogger.LogDebug($"{e}");
+            LogDebug("An error has occured while trying to create a debug log.", LogLevel.Warn, true);
+            LogDebug($"{e}");
         }
     }
-    
+
     public static void LogDebug(string input, LogLevel level = LogLevel.Debug, bool outputIfNotDebug = false)
     {
         if (_loaded)
         {
-            string log = $"[{level.ToString()}] {(!outputIfNotDebug ? "[Hidden] ": "")}" + input;
+            var log = $"[{level.ToString()}] {(!outputIfNotDebug ? "[Hidden] " : "")}" + input;
             if (!WriteDirectly)
-            {
                 Singleton._debugLogs.Add(log);
-            }
             else
-            {
                 File.AppendAllText(Singleton._filePath, "\n" + log);
-            }
         }
-            
-        if (outputIfNotDebug || DebugLogger.Debug)
-        {
+
+        if (outputIfNotDebug || Debug)
             switch (level)
             {
                 case LogLevel.Debug:
+#if EXILED
                     Log.Debug(input);
+#else
+                    Logger.Debug(input);
+#endif
                     break;
                 case LogLevel.Error:
+#if EXILED
                     Log.Error(input);
+#else
+                    Logger.Error(input);
+#endif
                     break;
                 case LogLevel.Info:
-                    Log.Info(input);
+#if EXILED
+                    Log.Warn(input);
+#else
+                    Logger.Warn(input);
+#endif
                     break;
                 case LogLevel.Warn:
-                    Log.Warn(input);
+#if EXILED
+                    Log.Info(input);
+#else
+                    Logger.Info(input);
+#endif
                     break;
             }
-        }
     }
 }
 
@@ -86,5 +96,5 @@ public enum LogLevel
     Debug = 0,
     Warn = 1,
     Error = 2,
-    Info = 3,
+    Info = 3
 }
