+++ AutoEvent/Plugin.cs
@@ -1,49 +1,77 @@
 ﻿using System;
 using System.IO;
-using HarmonyLib;
 using AutoEvent.API;
+using HarmonyLib;
+#if EXILED
 using Exiled.API.Features;
+#else
+using LabApi.Features;
+using LabApi.Features.Wrappers;
+using LabApi.Loader.Features.Paths;
+using LabApi.Loader.Features.Plugins;
+#endif
 
 namespace AutoEvent;
-public class AutoEvent : Plugin<Config>
+
+public class AutoEvent :
+#if EXILED
+    Plugin<Config>
+#else
+    Plugin<Config>
+#endif
 {
-    public override string Name => "AutoEvent";
-    public override string Author => "Created by a large community of programmers, map builders and just ordinary people, under the leadership of RisottoMan. MapEditorReborn for 14.1 port by Sakred_";
-    public override Version Version => Version.Parse("9.11.2");
-    public override Version RequiredExiledVersion => new(9, 6, 0);
-    public static string BaseConfigPath { get; set;}
+    public override string Name => "AutoEvent"; 
+
+    public override string Author =>
+        "Created by a large community of programmers, map builders and just ordinary people, under the leadership of RisottoMan. MapEditorReborn for 14.1 port by Sakred_. LabApi port by MedveMarci.";
+
+#if LABAPI
+    public override string Description => "A plugin that allows you to play mini-games in SCP:SL. It includes a variety of games such as Spleef, Lava, Hide and Seek, Knives, and more. Each game has its own unique mechanics and rules, providing a fun and engaging experience for players.";
+#endif
+
+    public override Version Version => Version.Parse("9.13.0");
+#if EXILED
+    public override Version RequiredExiledVersion => new(9, 8, 1);
+#else
+    public override Version RequiredApiVersion => new(LabApiProperties.CompiledVersion);
+#endif
+    public static string BaseConfigPath { get; set; }
     public static AutoEvent Singleton;
     public static Harmony HarmonyPatch;
     public static EventManager EventManager;
     private EventHandler _eventHandler;
-    
+
+#if EXILED
     public override void OnEnabled()
+#else
+    public override void Enable()
+#endif
     {
         if (!Config.IsEnabled) return;
 
         CosturaUtility.Initialize();
-        
+#if EXILED
         BaseConfigPath = Path.Combine(Paths.Configs, "AutoEvent");
-        
+#else
+        BaseConfigPath = Path.Combine(PathManager.Configs.FullName, "AutoEvent");
+#endif
         try
         {
             Singleton = this;
-            
+
             if (Config.IgnoredRoles.Contains(Config.LobbyRole))
             {
-                DebugLogger.LogDebug("The Lobby Role is also in ignored roles. This will break the game if not changed. The plugin will remove the lobby role from ignored roles.", LogLevel.Error, true);
+                DebugLogger.LogDebug(
+                    "The Lobby Role is also in ignored roles. This will break the game if not changed. The plugin will remove the lobby role from ignored roles.",
+                    LogLevel.Error, true);
                 Config.IgnoredRoles.Remove(Config.LobbyRole);
             }
 
             FriendlyFireSystem.IsFriendlyFireEnabledByDefault = Server.FriendlyFire;
 
-            var debugLogger = new DebugLogger(Config.AutoLogDebug);
             DebugLogger.Debug = Config.Debug;
-            if (DebugLogger.Debug)
-            {
-                DebugLogger.LogDebug($"Debug Mode Enabled", LogLevel.Info, true);
-            }
-            
+            if (DebugLogger.Debug) DebugLogger.LogDebug("Debug Mode Enabled", LogLevel.Info, true);
+
             try
             {
                 HarmonyPatch = new Harmony("autoevent");
@@ -64,7 +92,7 @@ public class AutoEvent : Plugin<Config>
                 CreateDirectoryIfNotExists(BaseConfigPath);
                 CreateDirectoryIfNotExists(Config.SchematicsDirectoryPath);
                 CreateDirectoryIfNotExists(Config.MusicDirectoryPath);
-                
+
                 // temporarily
                 DeleteDirectoryAndFiles(Path.Combine(BaseConfigPath, "Configs"));
                 DeleteDirectoryAndFiles(Path.Combine(BaseConfigPath, "Events"));
@@ -72,34 +100,32 @@ public class AutoEvent : Plugin<Config>
             }
             catch (Exception e)
             {
-                DebugLogger.LogDebug($"An error has occured while trying to initialize directories.", LogLevel.Warn, true);
+                DebugLogger.LogDebug("An error has occured while trying to initialize directories.", LogLevel.Warn,
+                    true);
                 DebugLogger.LogDebug($"{e}");
             }
 
             _eventHandler = new EventHandler(this);
             EventManager = new EventManager();
             EventManager.RegisterInternalEvents();
-            ConfigManager.LoadConfigsAndTranslations();
-            
-            DebugLogger.LogDebug($"The mini-games are loaded.", LogLevel.Info, true);
+
+            DebugLogger.LogDebug("The mini-games are loaded.", LogLevel.Info, true);
         }
         catch (Exception e)
         {
             DebugLogger.LogDebug("Caught an exception while starting plugin.", LogLevel.Warn, true);
             DebugLogger.LogDebug($"{e}");
         }
-        
+#if EXILED
         base.OnEnabled();
+#endif
     }
-    
+
     private static void CreateDirectoryIfNotExists(string path)
     {
         try
         {
-            if (!Directory.Exists(path))
-            {
-                Directory.CreateDirectory(path);
-            }
+            if (!Directory.Exists(path)) Directory.CreateDirectory(path);
         }
         catch (Exception e)
         {
@@ -107,15 +133,12 @@ public class AutoEvent : Plugin<Config>
             DebugLogger.LogDebug($"Path: {path}\n{e}");
         }
     }
-    
+
     private static void DeleteDirectoryAndFiles(string path)
     {
         try
         {
-            if (Directory.Exists(path))
-            {
-                Directory.Delete(path, true);
-            }
+            if (Directory.Exists(path)) Directory.Delete(path, true);
         }
         catch (Exception e)
         {
@@ -123,15 +146,20 @@ public class AutoEvent : Plugin<Config>
             DebugLogger.LogDebug($"Path: {path}\n{e}");
         }
     }
-    
+
+#if EXILED
     public override void OnDisabled()
+#else
+    public override void Disable()
+#endif
     {
         _eventHandler = null;
 
         HarmonyPatch.UnpatchAll();
         EventManager = null;
         Singleton = null;
-        
+#if EXILED
         base.OnDisabled();
+#endif
     }
 }
